+++ jest.config.web.js
@@ -1,8 +1,0 @@
-module.exports = {
-  testMatch: [
-    '<rootDir>/packages/**/web/**/*.test.js',
-    '<rootDir>/packages/**/web/**/*.test.jsx',
-  ],
-  setupFilesAfterEnv: ['./jest.setup.web.js'],
-  displayName: 'web',
-};