+++ packages/icons/src/index.native.js
@@ -1,0 +1,3 @@
+import { Star } from './svg';
+
+export { Star };